A sketch of a graph of segments of commits, along with visualization capabilities.
This serves as basis for algorithms that can produce such a data structure from a commit graph,
on the one side, and on the other, mappings to the current stack-based data structures.

These are in turn mapped to the current data structures as consumed by the UI.
