+++ Cargo.toml
@@ -76,6 +76,7 @@ but-workspace = { path = "crates/but-workspace" }
 but-hunk-assignment = { path = "crates/but-hunk-assignment" }
 but-hunk-dependency = { path = "crates/but-hunk-dependency" }
 but-db = { path = "crates/but-db" }
+but-graph = { path = "crates/but-graph" }
 but-action = { path = "crates/but-action" }
 but-status = { path = "crates/but-status" }
 git2-hooks = { version = "0.5.0" }
