+++ crates/but-graph/tests/graph/main.rs
@@ -1,0 +1,9 @@
+/// Tests for visualizing the graph data structure.
+mod vis {
+    use but_graph::Graph;
+
+    #[test]
+    fn basic() {
+        let _graph = Graph::default();
+    }
+}
