+++ crates/but-workspace/Cargo.toml
@@ -13,6 +13,7 @@ anyhow.workspace = true
 bstr.workspace = true
 git2.workspace = true
 but-core.workspace = true
+but-graph.workspace = true
 but-rebase.workspace = true
 gix = { workspace = true, features = ["worktree-mutation"] }
 gitbutler-stack.workspace = true
